Moving commands into their own folder and removing the error function
