+++ src/cli.ts
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 
 import { ls, run } from './index'
-import { ApplicationError, error } from './features/errors'
+import { ApplicationError } from './features/errors'
 
 async function cli (args: string[]) {
   const command = args.shift()
   if (command === 'ls') return ls()
   if (command === 'run' && args.length > 0) return run(args)
-  throw error('Invalid command')
+  throw new ApplicationError('Invalid command')
 }
 
 cli(process.argv.slice(2)).catch(error => {
